UPD: Massive Refactoring
