+++ tiatoolbox/cli.py
@@ -28,7 +28,7 @@ import numpy as np
 from PIL import Image
 
 from tiatoolbox import __version__, utils, wsicore
-from tiatoolbox.models.classification.patch_predictor import CNNPatchPredictor
+from tiatoolbox.models.controller.patch_predictor import CNNPatchPredictor
 from tiatoolbox.tools import stainnorm as sn
 from tiatoolbox.tools import tissuemask
 from tiatoolbox.utils.exceptions import MethodNotSupported
@@ -456,7 +456,7 @@ def tissue_mask(
     "available pretrained models please see "
     "https://tia-toolbox.readthedocs.io/en/latest/usage.html"
     "#tiatoolbox.models.classification.patch_predictor.get_pretrained_model",
-    default="resnet18-kather100K",
+    default="resnet18-kather100k",
 )
 @click.option(
     "--pretrained_weights",
