+++ tiatoolbox/models/architecture/utils.py
