+++ tiatoolbox/models/controller/patch_predictor.py
@@ -28,20 +28,17 @@ from typing import Callable, Tuple
 
 import numpy as np
 import torch
-import torch.nn as nn
 import tqdm
 
-from tiatoolbox import rcParam
-from tiatoolbox.models.abc import IOConfigABC, ModelABC
-from tiatoolbox.models.backbone import get_model
-from tiatoolbox.models.dataset import predefined_preproc_func
+from tiatoolbox.models.abc import IOConfigABC
+from tiatoolbox.models.architecture import get_pretrained_model
 from tiatoolbox.models.dataset.classification import PatchDataset, WSIPatchDataset
 from tiatoolbox.utils import misc
-from tiatoolbox.utils.misc import download_data, get_pretrained_model_info, save_as_json
+from tiatoolbox.utils.misc import save_as_json
 from tiatoolbox.wsicore.wsireader import VirtualWSIReader, get_wsireader
 
 
-class _IOConfigPatchPredictor(IOConfigABC):
+class IOPatchPredictorConfig(IOConfigABC):
     """Contain patch predictor input and output information."""
 
     # We pre-define to follow enforcement, actual initialization in init
@@ -58,85 +55,6 @@ class _IOConfigPatchPredictor(IOConfigABC):
         self.output_resolutions = output_resolutions
         for variable, value in kwargs.items():
             self.__setattr__(variable, value)
-
-
-class CNNPatchModel(ModelABC):
-    """Retrieve the model backbone and attach an extra FCN to perform classification.
-
-    Args:
-        backbone (str): Model name.
-        num_classes (int): Number of classes output by model.
-
-    Attributes:
-        num_classes (int): Number of classes output by the model.
-        feat_extract (nn.Module): Backbone CNN model.
-        pool (nn.Module): Type of pooling applied after feature extraction.
-        classifier (nn.Module): Linear classifier module used to map the features
-          to the output.
-
-    """
-
-    def __init__(self, backbone, num_classes=1):
-        super().__init__()
-        self.num_classes = num_classes
-
-        self.feat_extract = get_model(backbone)
-        self.pool = nn.AdaptiveAvgPool2d((1, 1))
-
-        # Best way to retrieve channel dynamically is passing a small forward pass
-        prev_num_ch = self.feat_extract(torch.rand([2, 3, 96, 96])).shape[1]
-        self.classifer = nn.Linear(prev_num_ch, num_classes)
-
-    # pylint: disable=W0221
-    # because abc is generic, this is actual definition
-    def forward(self, imgs):
-        """Pass input data through the model.
-
-        Args:
-            imgs (torch.Tensor): Model input.
-
-        """
-        feat = self.feat_extract(imgs)
-        gap_feat = self.pool(feat)
-        gap_feat = torch.flatten(gap_feat, 1)
-        logit = self.classifer(gap_feat)
-        prob = torch.softmax(logit, -1)
-        return prob
-
-    @staticmethod
-    def postproc(image):
-        """Define the post-processing of this class of model.
-
-        This simply applies argmax along last axis of the input.
-        """
-        return np.argmax(image, axis=-1)
-
-    @staticmethod
-    def infer_batch(model, batch_data, on_gpu):
-        """Run inference on an input batch.
-
-        Contains logic for forward operation as well as i/o aggregation.
-
-        Args:
-            model (nn.Module): PyTorch defined model.
-            batch_data (ndarray): A batch of data generated by
-              torch.utils.data.DataLoader.
-            on_gpu (bool): Whether to run inference on a GPU.
-
-        """
-        device = misc.select_device(on_gpu)
-
-        img_patches = batch_data
-        img_patches_device = img_patches.to(device).type(torch.float32)  # to NCHW
-        img_patches_device = img_patches_device.permute(0, 3, 1, 2).contiguous()
-
-        # Inference mode
-        model.eval()
-        # Do not compute the gradient (not training)
-        with torch.no_grad():
-            output = model(img_patches_device)
-        # Output should be a single tensor or scalar
-        return output.cpu().numpy()
 
 
 class CNNPatchPredictor:
@@ -530,7 +448,7 @@ class CNNPatchPredictor:
 
             self.iostate = self._iostate
             if patch_size is not None:
-                iostate = _IOConfigPatchPredictor(
+                iostate = IOPatchPredictorConfig(
                     input_resolutions=[{"resolution": resolution, "units": units}],
                     output_resolutions=[{"resolution": resolution, "units": units}],
                     patch_size=patch_size,
@@ -625,100 +543,3 @@ class CNNPatchPredictor:
             output = file_dict if len(imgs) > 1 or save_output else outputs
 
         return output
-
-
-def get_pretrained_model(pretrained_model=None, pretrained_weights=None):
-    """Load a predefined PyTorch model with the appropriate pretrained weights.
-
-    Args:
-        pretrained_model (str): Name of the existing models support by tiatoolbox
-          for processing the data. Currently supports:
-            - alexnet-kather100k: alexnet backbone trained on Kather 100k dataset.
-            - resnet18-kather100k: resnet18 backbone trained on Kather 100k dataset.
-            - resnet34-kather100k: resnet34 backbone trained on Kather 100k dataset.
-            - resnet50-kather100k: resnet50 backbone trained on Kather 100k dataset.
-            - resnet101-kather100k: resnet101 backbone trained on Kather 100k dataset.
-            - resnext5032x4d-kather100k: resnext50_32x4d backbone trained on Kather
-              100k dataset.
-            - resnext101_32x8d-kather100k: resnext101_32x8d backbone trained on
-              Kather 100k dataset.
-            - wide_resnet50_2-kather100k: wide_resnet50_2 backbone trained on
-              Kather 100k dataset.
-            - wide_resnet101_2-kather100k: wide_resnet101_2 backbone trained on
-              Kather 100k dataset.
-            - densenet121-kather100k: densenet121 backbone trained on
-              Kather 100k dataset.
-            - densenet161-kather100k: densenet161 backbone trained on
-              Kather 100k dataset.
-            - densenet169-kather100k: densenet169 backbone trained on
-              Kather 100k dataset.
-            - densenet201-kather100k: densenet201 backbone trained on
-              Kather 100k dataset.
-            - mobilenet_v2-kather100k: mobilenet_v2 backbone trained on
-              Kather 100k dataset.
-            - mobilenet_v3_large-kather100k: mobilenet_v3_large backbone trained on
-              Kather 100k dataset.
-            - mobilenet_v3_small-kather100k: mobilenet_v3_small backbone trained on
-              Kather 100k dataset.
-            - googlenet-kather100k: googlenet backbone trained on Kather 100k dataset.
-
-          By default, the corresponding pretrained weights will also be
-          downloaded. However, you can override with your own set of weights via
-          the `pretrained_weights` argument. Argument is case insensitive.
-        pretrained_weights (str): Path to the weight of the
-          corresponding `pretrained_model`.
-
-    Examples:
-        >>> # get mobilenet pretrained on kather by TIA team
-        >>> model = get_pretrained_model(pretrained_model='mobilenet_v2-kather100k')
-        >>> # get mobilenet defined by TIA team but loaded with user weight
-        >>> model = get_pretrained_model(
-        ...     pretrained_model='mobilenet_v2-kather100k'
-        ...     pretrained_weights='/A/B/C/my_weights.tar'
-        ... )
-
-    """
-    if not isinstance(pretrained_model, str):
-        raise ValueError("pretrained_model must be a string.")
-
-    # parsing protocol
-    pretrained_model = pretrained_model.lower()
-    backbone, dataset = pretrained_model.split("-")
-
-    pretrained_yml = get_pretrained_model_info()
-
-    pretrained_info = pretrained_yml[dataset]
-    pretrained_models_dict = pretrained_info["models"]
-
-    if backbone not in pretrained_models_dict.keys():
-        raise ValueError(f"Pretrained model `{pretrained_model}` does not exist.")
-
-    patch_size = pretrained_info["patch_size"]
-    resolution = pretrained_info["resolution"]
-    units = pretrained_info["units"]
-    num_classes = pretrained_info["num_classes"]
-
-    model = CNNPatchModel(backbone=backbone, num_classes=num_classes)
-    model.preproc_func = predefined_preproc_func(dataset)
-
-    if pretrained_weights is None:
-        pretrained_weights_url = pretrained_models_dict[backbone]
-        pretrained_weights_url_split = pretrained_weights_url.split("/")
-        pretrained_weights = os.path.join(
-            rcParam["TIATOOLBOX_HOME"], "models/", pretrained_weights_url_split[-1]
-        )
-        if not os.path.exists(pretrained_weights):
-            download_data(pretrained_weights_url, pretrained_weights)
-
-    # ! assume to be saved in single GPU mode
-    # always load on to the CPU
-    saved_state_dict = torch.load(pretrained_weights, map_location="cpu")
-    model.load_state_dict(saved_state_dict, strict=True)
-
-    iostate = _IOConfigPatchPredictor(
-        patch_size=patch_size,
-        stride_size=patch_size,
-        input_resolutions=[{"resolution": resolution, "units": units}],
-        output_resolutions=[{"resolution": resolution, "units": units}],
-    )
-    return model, iostate
