+++ tiatoolbox/models/controller/semantic_segmentor.py
